Separate raycast function
